requirements.txt
numpy>=1.20.0
matplotlib>=3.3.0
pandas>=1.3.0
pyyaml>=6.0
gitpython>=3.1.0
